Decouple navigator to a file
